Get rid off hardcoded node type names, insert commands refactored.
The refactoring of the insert commands means that the logic is now
way simpler, as our document assumptions have been simplified.

There was also a bug that we inserted 3 code blocks at a time (I think
because we were inserting on the edge of a node??), there is a TODO
there since I don't fully understand the cause of that bug yet.