Set correct themeStyle at initialization
Previously the default themeStyle was 'light' and the editor would
initialize the codemirror highlighting as if the theme was 'light'.
This would lead to a period where the theme was incorrectly set if the
user was in a dark style theme. Now we initialize the editor with the
correct theme style, without waiting for a message from the host.